check sesions before create an active sesion on a table

+++ backend/models/sesiones.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Sesiones struct {
-	ID         uint64     `json:"id" gorm:"primary_key"`
+	ID         uint     	`json:"id" gorm:"primary_key"`
 	Idmesa     uint       `json:"idmesa"`
 	Activo     bool       `json:"activo" gorm:"default:true"`
 	CreatedAt  time.Time  `json:"created_at"`
